# Tailwind config performance improvements