@EINazare 2021-05-17
Hey @pavelloz ,

Thank you for this PR.
Unfortunately, we cannot accept it, since the Purge CSS does not include the HTML files as well: https://github.com/creativetimofficial/notus-svelte/blob/main/public/index.html

As you can see inside our public/index.html we also have two classes here that would need to get compiled.

Best,
Manu